fix: fix python models
