+++ backend/python/pydevlake/pydevlake/__init__.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Any, Optional
+from typing import Optional
 
 import pytest
 pytest.register_assert_rewrite('pydevlake.testing')
