+++ backend/python/test/fakeplugin/fakeplugin/main.py
@@ -13,32 +13,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from enum import Enum
-from datetime import datetime
-from typing import Optional
 import json
+from datetime import datetime
 
-from pydantic import SecretStr
-
-from pydevlake import Plugin, Connection, Stream, ToolModel, ToolScope, ScopeConfig, RemoteScopeGroup, DomainType, \
-    Field, TestConnectionResult
+from fakeplugin.models import FakeConnection, FakeProject, FakeScopeConfig, FakePipeline
+from pydevlake import Plugin, Stream, RemoteScopeGroup, DomainType, TestConnectionResult
 from pydevlake.domain_layer.devops import CicdScope, CICDPipeline, CICDStatus, CICDResult, CICDType
-from pydevlake.migration import migration, MigrationScriptBuilder, Dialect
 
 VALID_TOKEN = "this_is_a_valid_token"
-
-
-class FakePipeline(ToolModel, table=True):
-    class State(Enum):
-        PENDING = "pending"
-        RUNNING = "running"
-        FAILURE = "failure"
-        SUCCESS = "success"
-
-    id: str = Field(primary_key=True)
-    started_at: Optional[datetime]
-    finished_at: Optional[datetime]
-    state: State
 
 
 class FakePipelineStream(Stream):
@@ -92,18 +74,6 @@ class FakePipelineStream(Stream):
                 finished_at=datetime(2023, 1, 10, 11, 3, 0, microsecond=i),
             ))
         return fake_pipelines
-
-
-class FakeConnection(Connection):
-    token: SecretStr
-
-
-class FakeProject(ToolScope, table=True):
-    url: str
-
-
-class FakeScopeConfig(ScopeConfig):
-    env: str
 
 
 class FakePlugin(Plugin):
@@ -167,12 +137,7 @@ class FakePlugin(Plugin):
         ]
 
 
-# test migration
-@migration(20230630000001, name="populated _raw_data_table column for fakeproject")
-def add_raw_data_params_table_to_scope(b: MigrationScriptBuilder):
-    b.execute(f'UPDATE {FakeProject.__tablename__} SET _raw_data_table = "_raw_fakeproject_scopes" WHERE 1=1', Dialect.MYSQL) #mysql only
-    b.execute(f'''UPDATE {FakeProject.__tablename__} SET _raw_data_table = '_raw_fakeproject_scopes' WHERE 1=1''', Dialect.POSTGRESQL) #mysql and postgres
-
+import fakeplugin.migrations  # NEEDED
 
 if __name__ == '__main__':
     FakePlugin.start()
